engine2
1 fix cannot snapshot problem(macro USE_SWSCALE)
2 fix webm ffextractor crash
3 build-in librtmp

MeetSDK
1 XOPlayer fix flv and mkv cannot seek problem
2 XOPlayer solve some ts cann't seek back problem when getTrackFormat()
3 XOPlayer fix ffextractor find_sps_pps() cannot find nalu in some ts media (after flv2ts convert)
4 XOPlayer fix ffextractor m_buffered_size < 0 bug(unsiged type mess-up)
5 XOPlayer suppport getMediaInfo()
6 support setDataSource(Context context, Uri uri, Map<String, String> headers) to set HTTP cookie
7 support GLES render(both HW_SYSTEM and SW mode) (still use native_window, much easily code)

foundation
1 sync foundation_rext code(fix h264_extradata_to_annexb)
2 update foundation build batch(support "enc mux librtmp openssl")

subtitle2
1 use libass SMP

+++ foundation/build/win32/config.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 USER_ROOT=`pwd`
-FFMPEG_HOME=$USER_ROOT/../../foundation_rext
+FFMPEG_HOME=$USER_ROOT/../../ffmpeg-2.8.3
 MS_INT_TYPES_HOME=$USER_ROOT/msinttypes
 PREFIX=$USER_ROOT/../../output/win32/$1
 
@@ -18,7 +18,7 @@ then
 echo "lite build"
 EXTRA_FF_BUILD_OPTION="\
 	--disable-decoders \
-	--enable-decoder=h263,h264,hevc,vp3,vp5,vp6,vp6a,vp6f,vp7,vp8,vp9,flv,mpegvideo,mpeg1video,mpeg2video,mpeg4,ac3,eac3,aac,dca,mp1,mp2,mp3,rv30,rv40,cook,wmv1,wmv2,wmv3,wmv3image,vorbis,ape,flac,wmav1,wmav2,wmapro,mjpeg,msmpeg4v1,msmpeg4v2,msmpeg4v3,tscc,gsm,gsm_ms,amrnb,amrwb,pcm_s16be,pcm_s16be_planar,pcm_s16le,pcm_s16le_planar,ass,dvbsub,dvdsub,mov_text,sami,srt,ssa,subrip,text \
+	--enable-decoder=h263,h264,hevc,vp3,vp5,vp6,vp6a,vp6f,vp7,vp8,vp9,flv,mpegvideo,mpeg1video,mpeg2video,mpeg4,ac3,eac3,aac,dca,mp1,mp2,mp3,rv30,rv40,cook,wmv1,wmv2,wmv3,wmv3image,vorbis,ape,flac,wmav1,wmav2,wmapro,mjpeg,msmpeg4v1,msmpeg4v2,msmpeg4v3,tscc,gsm,gsm_ms,amrnb,amrwb,pcm_s16be,pcm_s16be_planar,pcm_s16le,pcm_s16le_planar,ass,dvbsub,dvdsub,mov_text,sami,srt,ssa,subrip,text,h264_qsv,hevc_qsv,mpeg2_qsv,vc1_qsv \
 	--disable-demuxers \
 	--enable-demuxer=rm,mpegvideo,mjpeg,avi,h263,h264,hevc,matroska,dts,dtshd,aac,flv,mpegts,mpegps,mp4,m4v,mov,ape,hls,flac,rawvideo,realtext,rtsp,vc1,mp3,wav,asf,ogg \
 	--disable-parsers \
@@ -27,12 +27,11 @@ elif [ ${1}x == 'micro'x ]; then
 echo "micro build"
 EXTRA_FF_BUILD_OPTION="\
 	--disable-decoders \
-	--enable-decoder=h264,aac \
+	--enable-decoder=h264,aac,h264_qsv,hevc_qsv,mpeg2_qsv,vc1_qsv \
 	--disable-demuxers \
 	--enable-demuxer=h264,mp4,mov,mpegts,flv,hls \
 	--disable-parsers \
 	--enable-parser=h264,aac_latm \
-	--enable-muxer=mpegts,flv,hls \
 	--disable-protocols \
 	--enable-protocol=file,http,rtmp,hls \
 	--disable-filters \
@@ -76,33 +75,41 @@ fi
 
 #	--enable-decoder=aac_latm
 
-rm -rf $PREFIX
+if [ -d $PREFIX ]; then
+	rm -rf $PREFIX
+fi
 
 #--enable-demuxer=rm,mpegvideo,avi,h263,h264,hevc,mkv,aac,flv,mpegts,mp4,mov,ape,hls,rtsp,rtp,sdp,asf \
 
 echo "run configure"
 ./configure --prefix=$PREFIX \
-	--toolchain=icl \
+	--toolchain=msvc \
 	--arch=i686 \
     --enable-shared \
     --disable-static \
 	--enable-w32threads \
 	--enable-optimizations \
-	--extra-cflags="-DNDEBUG -march=i686 -Z7 -DX264_API_IMPORTS -I${MS_INT_TYPES_HOME}" \
+	--extra-cflags="-DNDEBUG -march=i686 -Z7 -DX264_API_IMPORTS -I${MS_INT_TYPES_HOME} -I../thirdparty/mfx/include" \
 	--extra-ldflags="-DEBUG" \
 	--optflags="-O2" \
 	--disable-doc \
 	--disable-avdevice \
 	--disable-dxva2 \
 	--disable-swscale-alpha \
-	--disable-encoders \
+	--enable-encoder=h264_qsv,hevc_qsv,mpeg2_qsv,vc1_qsv \
 	--disable-muxers \
+	--enable-muxer=flv,mpegts,mp4,webm \
 	--disable-devices \
-	--disable-asm \
+	--enable-libmfx \
 	--disable-symver $EXTRA_FF_BUILD_OPTION
+	
+# dosomething1 || exit 1	
+if [[ $? -ne 0 ]]; then
+	echo -e "\033[31m \n\nfailed to config ffmpeg \033[0m"
+    exit 1
+fi
 
 echo "begin to building..."
-read n
 
 # delete old files
 make clean
